Добавил автомаппинг

+++ src/BulletinBoard/Application/BulletinBoard.AppServices/Contexts/Announcements/Services/AnnouncementService.cs
@@ -1,4 +1,6 @@
+using AutoMapper;
 using BulletinBoard.AppServices.Contexts.Announcements.Repositories;
+using BulletinBoard.AppServices.Exceptions;
 using BulletinBoard.Contracts.Announcements.Requests;
 using BulletinBoard.Contracts.Announcements.Responses;
 using BulletinBoard.Domain.Entities;
@@ -7,7 +9,8 @@ namespace BulletinBoard.AppServices.Contexts.Announcements.Services;
 
 /// <inheritdoc />
 public class AnnouncementService(
-    IAnnouncementRepository announcementRepository
+    IAnnouncementRepository announcementRepository,
+    IMapper mapper
 ) : IAnnouncementService
 {
     /// <inheritdoc />
@@ -15,79 +18,47 @@ public class AnnouncementService(
         CancellationToken cancellationToken)
     {
         var announcements = await announcementRepository.GetByFilterAsync(filter, cancellationToken);
-
-        var response = announcements.Select(announcement => new AnnouncementResponse
-        {
-            Id = announcement.Id,
-            Title = announcement.Title,
-            Description = announcement.Description,
-            CreatedAt = announcement.CreatedAt
-        }).ToList().AsReadOnly();
-
-        return response;
+        return mapper.Map<IReadOnlyCollection<AnnouncementResponse>>(announcements);
     }
 
     /// <inheritdoc />
     public async Task<AnnouncementResponse> GetByIdAsync(Guid announcementId, CancellationToken cancellationToken)
     {
-        var announcement = await announcementRepository.GetByIdAsync(announcementId, cancellationToken);
-
-        var response = new AnnouncementResponse
-        {
-            Id = announcement.Id,
-            Title = announcement.Title,
-            Description = announcement.Description,
-            CreatedAt = announcement.CreatedAt
-        };
-
-        return response;
+        var announcement = await GetAnnouncementOrThrowAsync(announcementId, cancellationToken);
+        return mapper.Map<AnnouncementResponse>(announcement);
     }
 
     /// <inheritdoc />
-    public async Task<Guid> CreateAsync(CreateAnnouncementRequest announcementRequest,
+    public async Task<Guid> CreateAsync(CreateAnnouncementRequest request,
         CancellationToken cancellationToken)
     {
-        var announcement = new Announcement
-        {
-            Id = Guid.NewGuid(),
-            CreatedAt = DateTime.UtcNow,
-            Title = announcementRequest.Title,
-            Description = announcementRequest.Description
-        };
-
-        var result = await announcementRepository.CreateAsync(announcement, cancellationToken);
-
-        return result;
+        var announcement = mapper.Map<CreateAnnouncementRequest, Announcement>(request);
+        announcement.CreatedAt = DateTime.UtcNow;
+        return await announcementRepository.CreateAsync(announcement, cancellationToken);
     }
 
     /// <inheritdoc />
-    public async Task<AnnouncementResponse> UpdateAsync(Guid announcementId,
-        UpdateAnnouncementRequest announcementRequest, CancellationToken cancellationToken)
+    public async Task<AnnouncementResponse> UpdateAsync(Guid announcementId, UpdateAnnouncementRequest request, CancellationToken cancellationToken)
     {
-        var announcement = new Announcement
-        {
-            Id = announcementId,
-            Title = announcementRequest.Title,
-            Description = announcementRequest.Description
-        };
-
-        var updatedAnnouncement =
-            await announcementRepository.UpdateAsync(announcementId, announcement, cancellationToken);
-
-        var response = new AnnouncementResponse
-        {
-            Id = updatedAnnouncement.Id,
-            Title = updatedAnnouncement.Title,
-            Description = updatedAnnouncement.Description,
-            CreatedAt = updatedAnnouncement.CreatedAt
-        };
-
-        return response;
+        var announcement = await GetAnnouncementOrThrowAsync(announcementId, cancellationToken);
+        mapper.Map(request, announcement);
+        await announcementRepository.UpdateAsync(announcement, cancellationToken);
+        return mapper.Map<AnnouncementResponse>(announcement);
     }
-
+    
     /// <inheritdoc />
     public async Task DeleteAsync(Guid announcementId, CancellationToken cancellationToken)
     {
+        var exists = await announcementRepository.ExistsAsync(announcementId, cancellationToken);
+        if (!exists) throw new NotFoundException(announcementId.ToString());
         await announcementRepository.DeleteAsync(announcementId, cancellationToken);
+    }
+    
+    
+    
+    private async Task<Announcement> GetAnnouncementOrThrowAsync(Guid announcementId, CancellationToken cancellationToken)
+    {
+        var announcement = await announcementRepository.GetByIdAsync(announcementId, cancellationToken);
+        return announcement ?? throw new NotFoundException(announcementId.ToString());
     }
 }
